requirements.txt
openai
pdfplumber
faiss-cpu
tiktoken
python-dotenv
fastapi
uvicorn
numpy